+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change log
 
+## Unreleased
+
 ## 0.4.2-test - 2023-09-26
 
 ### Added