+++ CHANGELOG.md
@@ -152,3 +152,4 @@
 ## 0.0.1 - 2017-01-26
 
 Initial release.
+bla